+++ libraries/MCP3424/MCP3424.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: MCP3424.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.5
+// VERSION: 0.2.0
 // PURPOSE: Arduino library for 18 bit ADC I2C MCP3424 and compatibles.
 //     URL: https://github.com/RobTillaart/MCP3424
 
@@ -24,14 +24,23 @@ MCP3424::MCP3424(uint8_t address, TwoWire *wire)
   _gain = 1;
   _bits = 12;
   _config = 0x10;  //  default
+  _lastRead = 0;
+  _raw = 0;
 }
-
 
 bool MCP3424::begin()
 {
-  return isConnected();
+   //  set defaults
+  _gain = 1;
+  _bits = 12;
+  _config = 0x10;
+  if (isConnected())
+  {
+    writeConfig();
+    return true;
+  }
+  return false;
 }
-
 
 bool MCP3424::isConnected()
 {
@@ -39,22 +48,35 @@ bool MCP3424::isConnected()
   return ( _wire->endTransmission() == 0);
 }
 
-
 uint8_t MCP3424::getAddress()
 {
   return _address;
 }
-
 
 uint8_t MCP3424::getMaxChannels()
 {
   return _maxChannels;
 }
 
-
 int32_t MCP3424::read()
 {
   return readRaw();
+}
+
+void MCP3424::requestSingleShot()
+{
+  setSingleShotMode();
+}
+
+bool MCP3424::isReady()
+{
+  readRaw();
+  return (_config & 0x80) == 0x00;
+}
+
+uint32_t MCP3424::lastRead()
+{
+  return _lastRead;
 }
 
 
@@ -70,12 +92,10 @@ float MCP3424::readVolts()
   return read() * (15.625e-6 * (1L << (18 - _bits))) / _gain;
 }
 
-
 float MCP3424::readMilliVolts()
 {
   return read() * (15.625e-3 * (1L << (18 - _bits))) / _gain;
 }
-
 
 float MCP3424::readMicroVolts()
 {
@@ -103,7 +123,6 @@ bool MCP3424::setChannel(uint8_t channel)
   return true;
 }
 
-
 uint8_t MCP3424::getChannel()
 {
   return _channel;
@@ -120,7 +139,9 @@ bool MCP3424::setGain(uint8_t gain)
   if (_gain != gain)
   {
     _gain = gain;
-    _config &= 0xFC;  //  gain == 1
+    //  clear gain bits ==> gain == 1
+    _config &= 0xFC;
+    //  set gain bits for gain != 1
     if (_gain == 2)      _config |= 0x01;
     else if (_gain == 4) _config |= 0x02;
     else if (_gain == 8) _config |= 0x03;
@@ -128,7 +149,6 @@ bool MCP3424::setGain(uint8_t gain)
   }
   return true;
 }
-
 
 uint8_t MCP3424::getGain()
 {
@@ -155,7 +175,6 @@ bool MCP3424::setResolution(uint8_t bits)
   return true;
 }
 
-
 uint8_t MCP3424::getResolution()
 {
   //  return 12 + 2 * ((_config >> 2) & 0x03);
@@ -170,7 +189,6 @@ uint16_t MCP3424::getConversionDelay()
   return _interval[idx];
 }
 
-
 void MCP3424::setContinuousMode()
 {
   if (getMode() != 1)
@@ -183,11 +201,9 @@ void MCP3424::setContinuousMode()
 
 void MCP3424::setSingleShotMode()
 {
-  if (getMode() != 0)
-  {
-    _config &= ~0x10;
-    writeConfig();
-  }
+  _config &= ~0x10;  //  set single shot mode
+  _config |= 0x80;   //  and trigger single shot
+  writeConfig();
 }
 
 
@@ -209,6 +225,49 @@ int MCP3424::writeConfig()
 }
 
 
+/*
+//  Optimized readRaw, need test with hardware.
+
+//  status bit/byte is not read.
+int32_t MCP3424::readRaw()
+{
+  int32_t rv = 0;
+
+  uint8_t bitsToRead = 3;
+  if (_bits == 18) bitsToRead = 4;
+  uint32_t signBit = (1UL << (_bits -1));
+  uint32_t extendSign = ~(signBit -1);
+
+  if (_wire->requestFrom(_address, (uint8_t)bitsToRead) != bitsToRead)
+  {
+    //  _error = MCP3424_I2C_ERROR;
+    return 0;
+  }
+  //  read data bytes
+  if (bitsToRead == 4)
+  {
+    rv += _wire->read();
+    rv <<= 8;
+  }
+  rv += _wire->read();
+  rv <<= 8;
+  rv += _wire->read();
+
+  //  read configuration byte
+  _config = _wire->read();
+  //  handle sign bit.
+  if (rv & signBit)
+  {
+    //  sign extend 14 bits
+    rv |= extendSign;
+  }
+  _lastRead = millis();
+  _raw = rv;
+  return rv;
+}
+*/
+
+
 //  status bit/byte is not read.
 int32_t MCP3424::readRaw()
 {
@@ -217,31 +276,60 @@ int32_t MCP3424::readRaw()
   //  not optimized compacted code.
   if (_bits == 18)
   {
-    if (_wire->requestFrom(_address, (uint8_t)3) != 3)
+    if (_wire->requestFrom(_address, (uint8_t)4) != 4)
     {
       //  _error = MCP3424_I2C_ERROR;
       return 0;
     }
+    //  read data bytes
     rv += _wire->read();
     rv <<= 8;
     rv += _wire->read();
     rv <<= 8;
     rv += _wire->read();
+    //  read the configuration byte
+    _config = _wire->read();
     //  handle sign bit.
-    if (rv & 0x00020000) rv |= 0xFFFC0000;
+    if (rv & 0x00020000)
+    {
+      //  sign extend 14 bits
+      rv |= 0xFFFC0000;
+    }
+    _lastRead = millis();
+    _raw = rv;
     return rv;
   }
 
   //  12, 14, 16 bit
-  if (_wire->requestFrom(_address, (uint8_t)2) != 2)
+  if (_wire->requestFrom(_address, (uint8_t)3) != 3)
   {
     //  _error = MCP3424_I2C_ERROR;
     return 0;
   }
+  //  read data bytes
   rv += _wire->read();
   rv <<= 8;
   rv += _wire->read();
-  //  handle sign bit not needed.
+  //  read the configuration byte
+  _config = _wire->read();
+  //  handle sign bit
+  if ((_bits == 12) && (rv & 0x0800))
+  {
+    //  sign extend 20 bits
+    rv |= 0xFFFFF000;
+  }
+  if ((_bits == 14) && (rv & 0x2000))
+  {
+    //  sign extend 18 bits
+    rv |= 0xFFFFC000;
+  }
+  if ((_bits == 16) && (rv & 0x8000))
+  {
+    //  sign extend 16 bits
+    rv |= 0xFFFF0000;
+  }
+  _lastRead = millis();
+  _raw = rv;
   return rv;
 }
 
@@ -264,6 +352,11 @@ MCP3422::MCP3422(uint8_t address, TwoWire *wire) : MCP3424(address, wire)
 MCP3423::MCP3423(uint8_t address, TwoWire *wire) : MCP3424(address, wire)
 {
   _maxChannels = 2;
+}
+
+MCP3425::MCP3425(uint8_t address, TwoWire *wire) : MCP3424(address, wire)
+{
+  _maxChannels = 1;
 }
 
 MCP3426::MCP3426(uint8_t address, TwoWire *wire) : MCP3424(address, wire)